+++ UI17/src/locale/i18nStrings.ts
@@ -1,4 +1,0 @@
-export const i18nStrings = {
-  de: {},
-  en: {},
-};