+++ frontend/src/app/menu/main-menu/main-menu.component.ts
@@ -1,0 +1,9 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-main-menu',
+    templateUrl: './main-menu.component.html',
+    styleUrls: ['./main-menu.component.scss'],
+    standalone: false
+})
+export class MainMenuComponent {}