+++ frontend/src/app/modeless/flexible-area/flexible-area.component.ts
@@ -1,0 +1,9 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-flexible-area',
+    templateUrl: './flexible-area.component.html',
+    styleUrls: ['./flexible-area.component.scss'],
+    standalone: false
+})
+export class FlexibleAreaComponent {}