+++ frontend/src/app/mud/files.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FilesService } from './files.service';
+
+describe('FilesService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: FilesService = TestBed.inject(FilesService);
+    expect(service).toBeTruthy();
+  });
+});