+++ VERSION_HISTORY.md
@@ -37,3 +37,6 @@
 37. Version 0.3.0: Lots of Bugfixes, minor version upgrades, Inventory Display. no PWA yet!
 38. Version 0.4.0: Lots of Bugfixes, minor version upgrades
 39. Version 0.5.0: Comndensed configuration into one docker image and into main Dockerfile
+40. Version 0.6.0: update from angular,primeng,etc from v17 to v19 and got it up and running on localhost.
+41. Version 0.6.1: fixed DynamicDialogs,disabled pwa,reduced logs,except Editor
+42. Version 0.6.2: more log fixes in backend and frontend, removed i18n